--- a/tonic/src/codec/mod.rs
+++ b/tonic/src/codec/mod.rs
@@ -3,7 +3,6 @@
 //! This module contains the generic `Codec`, `Encoder` and `Decoder` traits
 //! and a protobuf codec based on prost.
 
-mod buffer;
 #[cfg(feature = "compression")]
 pub(crate) mod compression;
 mod decode;
@@ -12,11 +11,9 @@ mod encode;
 mod prost;
 
 use crate::Status;
-use std::io;
 
 pub(crate) use self::encode::{encode_client, encode_server};
 
-pub use self::buffer::{DecodeBuf, EncodeBuf};
 #[cfg(feature = "compression")]
 #[cfg_attr(docsrs, doc(cfg(feature = "compression")))]
 pub use self::compression::{CompressionEncoding, EnabledCompressionEncodings};
@@ -40,7 +37,7 @@ pub trait Codec: Default {
     type Decode: Send + 'static;
 
     /// The encoder that can encode a message.
-    type Encoder: Encoder<Item = Self::Encode, Error = Status> + Send + Sync + 'static;
+    type Encoder: Encoder<Self::Encode, Error = Status> + Send + Sync + 'static;
     /// The encoder that can decode a message.
     type Decoder: Decoder<Item = Self::Decode, Error = Status> + Send + Sync + 'static;
 
@@ -50,32 +47,4 @@ pub trait Codec: Default {
     fn decoder(&mut self) -> Self::Decoder;
 }
 
-/// Encodes gRPC message types
-pub trait Encoder {
-    /// The type that is encoded.
-    type Item;
-
-    /// The type of encoding errors.
-    ///
-    /// The type of unrecoverable frame encoding errors.
-    type Error: From<io::Error>;
-
-    /// Encodes a message into the provided buffer.
-    fn encode(&mut self, item: Self::Item, dst: &mut EncodeBuf<'_>) -> Result<(), Self::Error>;
-}
-
-/// Decodes gRPC message types
-pub trait Decoder {
-    /// The type that is decoded.
-    type Item;
-
-    /// The type of unrecoverable frame decoding errors.
-    type Error: From<io::Error>;
-
-    /// Decode a message from the buffer.
-    ///
-    /// The buffer will contain exactly the bytes of a full message. There
-    /// is no need to get the length from the bytes, gRPC framing is handled
-    /// for you.
-    fn decode(&mut self, src: &mut DecodeBuf<'_>) -> Result<Option<Self::Item>, Self::Error>;
-}
+pub use tokio_util::codec::{Decoder, Encoder};