@LucioFranco 2022-02-15
We will be keeping the tonic codec traits for versioning purposes. Thanks for the contribution!